moving fetcher logic to separate container
