+++ backend/main.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/jackc/pgx/v5"
-	"github.com/redis/go-redis/v9"
 )
 
 // todo; upgrade to structured logging
@@ -50,9 +49,6 @@ func run() error {
 	// Setting up routes in a separate file
 	setupRoutes(r)
 
-	//
-	ExampleClient()
-
 	fmt.Println("Server is running on port:", port)
 	if err := http.ListenAndServe(":"+port, r); err != nil {
 		return err
@@ -69,37 +65,4 @@ func dbConnect(ctx context.Context) (*api.Queries, error) {
 
 	queries := api.New(conn)
 	return queries, nil
-}
-
-func ExampleClient() {
-
-	var ctx = context.Background()
-
-	rdb := redis.NewClient(&redis.Options{
-		Addr:     "redis:6379",
-		Password: "", // no password set
-		DB:       0,  // use default DB
-	})
-
-	err := rdb.Set(ctx, "key", "value", 0).Err()
-	if err != nil {
-		panic(err)
-	}
-
-	val, err := rdb.Get(ctx, "key").Result()
-	if err != nil {
-		panic(err)
-	}
-	fmt.Println("key", val)
-
-	val2, err := rdb.Get(ctx, "key2").Result()
-	if err == redis.Nil {
-		fmt.Println("key2 does not exist")
-	} else if err != nil {
-		panic(err)
-	} else {
-		fmt.Println("key2", val2)
-	}
-	// Output: key value
-	// key2 does not exist
 }
