+++ fetcher/fetcher.dockerfile
@@ -1,0 +1,12 @@
+FROM golang:alpine
+
+WORKDIR /app
+
+COPY go.mod go.sum ./
+RUN go mod download
+
+COPY . .
+
+RUN go build -o fetcher .
+
+CMD ["./fetcher"]
