+++ fetcher/go.mod
@@ -1,0 +1,10 @@
+module danothebard.com/m/v2
+
+go 1.22.4
+
+require github.com/redis/go-redis/v9 v9.5.3
+
+require (
+	github.com/cespare/xxhash/v2 v2.2.0 // indirect
+	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
+)
